user_function

+++ app/Controllers/Dashboard/FunctionController.php
@@ -7,15 +7,21 @@ use App\Models\FunctionModel;
 class FunctionController extends BaseController
 {
     private $function_model;
+    private $uf = array();
     public function __construct()
     {
         $this->function_model = new FunctionModel();
+
     }
 
     public function index()
     {
-        $meta = array('page_title' => lang('AppLang.page_title_functions'));
-        return $this->page_construct('dashboard/function_view', $meta);
+        $this->uf = $this->libauth->getFunction('function');
+        if($this->uf['view']) {
+            $meta = array('page_title' => lang('AppLang.page_title_functions'));
+            return $this->page_construct('dashboard/function_view', $meta);
+        }else
+            return view('errors/html/error_403');
     }
     public function function_ajax()
     {
@@ -27,32 +33,36 @@ class FunctionController extends BaseController
     }
     public function add_function()
     {
-        if($this->request->getPost())
+        if($this->request->getPost()&&$this->uf['add'])
         {
             $data_function = $this->request->getPost();
             $data['result'] = ($this->function_model->add_function($data_function));
             $data['message']= $this->function_model->get_messages();
             echo json_encode(array_values($data));
         }
+        else
+            echo json_encode(array_values($this->libauth->getError()));
     }
     public function edit_function()
     {
-        if($this->request->getPost())
+        if($this->request->getPost()&&$this->uf['edit'])
         {
             $data_function = $this->request->getPost();
             $data['result'] = ($this->function_model->edit_function($data_function));
             $data['message']= $this->function_model->get_messages();
             echo json_encode(array_values($data));
-        }
+        }else
+            echo json_encode(array_values($this->libauth->getError()));
     }
     public function delete_function()
     {
-        if($this->request->getPost())
+        if($this->request->getPost()&&$this->uf['delete'])
         {
             $data_function = $this->request->getPost();
             $data['result'] = ($this->function_model->delete_function($data_function));
             $data['message']= $this->function_model->get_messages();
             echo json_encode(array_values($data));
-        }
+        }else
+            echo json_encode(array_values($this->libauth->getError()));
     }
 }
